Make <meta content=foo property=og:blah> work.
Fixes W3C bug #16582.
https://www.w3.org/Bugs/Public/show_bug.cgi?id=16582

The schema treated the meta element with content and property
attributes (as used by, e.g., Facebook Open Graph markup) as valid
in RDFa (full) but not for RDFa Lite. The RDFa content attribute
is part of RDFa full but not RDFa Lite. So we need to piggyback
off meta.name.elem, which already allows the content attribute,
and just require there either the name attribute OR the property
attribute if the content attribute is present, or (as already
specified in other parts of the schema), http-equiv or itemprop.