Update Html5SpecBuilder to match current spec. r=hsivonen
This updates he spec builder to handle a change to the spec that
added hyperlinks within some <dt> elements we parse for; e.g.:

<dt><a href=#element-dfn-attributes>Content attributes</a>:</dt>

Because this code change doesn't actually add a new start for the
<a> element within <dt>, it doesn't allow dealing with text nodes
in <dt> the way that it really should. Instead it adds a hack that
causes the characters method to just ignore those text nodes.
Without that hack, the colon beofre the </dt> end tag gets
undesireably read into the spec fragments and shows up in the UI.