Allow empty string for all RDFa attribute values.
Act of faith. The RDFa 1.1 Core spec contains a non-normative Note
claiming that 'In all cases it is possible for these attributes to
be used with no value (e.g., @datatype="")', though nowhere in the
normative conformance requirements of the spec is that actually
specified. So I guess we take their word for it.