Make ImageReportAdviceBuilder deal with current wiki markup.
The markup generated by the backend of wiki.whatwg.org apparently
changed at some point recently. It now has a
<div id="mw-content-text"> wrapper around the content where it
didn't have one before, and the existing ImageReportAdviceBuilder
didn't seem to be expecting that and ended up failing with an NPE.

This change fixes it. It may not be the best way to fix it but
given that ImageReportAdviceBuilder is pretty fragile anyway, it
doesn't seem worth trying to finesse it right now.