Make HTML5+ARIA schema conform with spec. b=833
This patch is not yet complete in that it doesn't yet bring the
ARIA support fully into conformance with the current spec. But
better to land it now rather than waiting another year to try to
complete it. We can of course make updates to the ARIA support in
response to specific bug reports. And later I can get back to
making incremental updates to it.