Send Accept header that matches typical browser.
This change causes us to always send the following Accept header:

  text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8

That header value matches what a typical browser sends for Accept.
We send it regardless of what type of Preset or Parser option the
user has selected. The benefit is that for any given document a
user validates, the validator will get the same version of that
document that browsers get.