Drop warnings about "cite" attr per current spec.
The current spec says, the cite attribute is "primarily intended
for private use (e.g. by server-side scripts collecting statistics
about a site's use of quotations), not for readers".

So since the "The cite attribute on the blockquote element is not
supported by browsers yet" warning message doesn't fit with the
guidance in the spec, we shouldn't keep emitting the warning.