Upgraded to Jetty 9.2.9
Support for AJP13 has been dropped with Jetty 9, so as a consequence of
this upgrade, we lose AJP as an option for the checker service. But the
Jetty wiki says, "We do NOT recommend AJP. Use HTTP and mod_proxy instead."
so it sounds like we're probably better off without the AJP option anyway.
https://wiki.eclipse.org/Jetty/Howto/Configure_AJP13