Work around broken `id` checking for XHTML in CLI
When using the command-line interface, it seems 456abf0 (for showing
extracts, issue #53) regressed the ability of the XML parser to properly
assign ID-ness to id attributes. I've opened #72 for that and need to
investigate further. For now, this change works around the one place I've
found so far that's affected by the regression.