Exit the meta[name]-checking business.
Removed the code for checking if meta[name] values are registered. While
admiring the spirit behind the original requirement, the experiment has
failed in practice. It's ended up costing a huge amount of maintenance time
for something that was of very little benefit to end users to begin with --
even if it had ever worked as conceived. But it's not working as conceived
-- in fact at this point it's mostly just wasting the time of both users
and maintainers. So, enough is enough -- let's cut our losses.