Fix display of images in Image Report feature
This fixes a major bug in the Image Report feature that resulted in all
images with relative URLs failing to actually be displayed (making the
feature useless in cases where you relative URLs for images in your docs).

The bug was due to the fact that the checker code for resolving relative
URLs was closely bound to the parsing API of Jena IRI library, so when we
switched the checker to relying on Galimatias instead of that Jena IRI
library, all our handling of relative URLs was silently regressed/broken.

Fixes https://github.com/validator/validator/issues/245. Thanks @ikar-us.