Experiment with dropping boilerplate info messages
In the interest of making the least amount of noise for users in the normal
case where someone is just checking an HTML document, this change suppresses
the following info-level messages that the checker has always emitted

  - “The Content-Type was text/html. Using the HTML parser.”
  - “Using the schema for HTML5 + SVG 1.1 + MathML 3.0 + RDFa Lite 1.1.”

Instead for that normal case, the content-type and schema details are shown
only in the final success/failure message after the warnings/errors (if any).

For the non-normal case where someone is using a non-text/html content type
and/or a schema other than the default HTML schema, the checker continues
to emit info messages with the content-type and schema details.

Relates to https://github.com/validator/validator/issues/239#issuecomment-187855741