Check for missing lang and mislabeled lang
Fixes https://github.com/validator/validator/issues/284

This runs https://github.com/shuyo/language-detection on every document
checked, and if the document has an `html[lang]` value that does not match
the detected language, a new warning is emitted; and a related warning is
emitted if the document has no `html[lang]` at all.

The goal is to help authors avoid the mistake of, for example, having a
document labeled `html[lang=en]` whose actual contents are actually in a
language other than English. Such lang-mislabeled documents cause
particular problems for screen-reader users; for some background, see:

  https://www.w3.org/Bugs/Public/show_bug.cgi?id=26942#c5

The implementation of this feature is highly experimental and may need to
be reverted (and maybe rewritten) if the performance impact is too high. It
increases the RAM usage of the checker non-insignificantly; for checking
large multi-megabyte docs, it requires setting the Java heap size somewhere
above 128MB to avoid OOM errors from the Java GC.

It introduces 3 more dependencies: https://github.com/shuyo/language-detection
for the actual language detection, and the Apache Stanbol Language Detection
convenience wrapper around that, along with https://jsonic.osdn.jp/.