Ignore cookie headers in responses
Tons of real-world documents on the Web have cookie headers that are not
strictly RFC-conformant. The Apache HttpClient library that we use causes
error messages to be logged for every such document, resulting in unwelcome
noise in our log output. We do nothing with the cookies anyway, so let’s
try completely ignoring all cookies.