Stop throwing for “HTTP resource not retrievable”
The checker service relatively commonly receives HTTP errors for requests
to check particular documents, because the response comes back 404 or 500
or whatever due to some error on the server where the document is hosted.

Every one of those cases results in a stack trace getting output to the log
for the checker service, resulting in unnecessary noise in the log.

The service also emits the HTTP error to the user in the checker Web UI or
in the JSON/GNU extra output from the request through the checker REST API.
So logging the error to the log for the checker service is redundant anyway.

This change prevents those errors from being output to checker-service log.