Don’t log stack trace for SAXException
org.eclipse.jetty.io.EofException seems to be the only case where a
SAXException ends up getting logged (along with an even huger-than-usual
stack trace). There is no usefulness in logging a stack trace for that, so
let’s instead just log a one-line message and not the stack trace.