For XML input only do language detection for XHTML
This works around a problem that causes checking to fail for the Citation Styles
checker at http://validator.citationstyles.org/ Thanks @rmzelle for reporting

The cause is that language-detection checker wraps an XMLReader, and if that
language-detecting wrapper is used when the schema is not XHTML5, it apparently
breaks some assumption the language-detecting wrapper is (unintentionally)
making, and the call to endDocument() in the content handler and ends up causing
an NPE in net.sf.saxon.tinytree.TinyBuilder.close(TinyBuilder.java:176).

The language-detection feature is really only intended for use with (X)HTML(5),
so this change works around the problem by only using the language-detection
wrapper if the schema that’s been selected is the XHTML5 schema.