Don’t include <a> content in langdetect
The language detector seems to misidentify the language of documents
with text that is mostly in hyperlinks (e.g., product pages) more often
than it does with documents that are actual prose. Not including the
content of hyperlinks in the text we run the language detector on
reduces the frequency of those kinds of misidentifications.