Don’t include <td> content in langdetect
The language detector seems to misidentify the language of documents
whose main content tables with a lot of non-prose more often than it
does documents that are mostly actual prose. Not including the content
of table cells in the text we run the language detector on reduces the
frequency of those kinds of misidentifications.