Make CLI use the right schema for .xhtml files
This causes the CLI load the XML flavor of the HTML schema (RelaxNG
grammar) the first time it sees a .xhtml file, and then not do a schema
reload again unless/until it’s given an .html file (in which case it
switches back to the default text/html flavor of the HTML schema).

In the case where you’re using the CLI to check a large number of .xhtml
and .html files together during the same run, this change may cause the
CLI to take noticeably longer to check the files.

If you do notice the CLI taking longer, try running it with the `--html`
option, to force it to always use the HTML parser and the text/html
schema flavor (and never use the XML parser or XML schema flavor).

Fixes https://github.com/validator/validator/issues/415 Thanks @caesar