Make checker service ignore SSL errors by default
This changes the default behavior of the Web-based checker service for
handling of sites with TLS/SSL problems. Previously the default behavior
was for the checker service to refuse to fetch a document for checking
if it finds any problems in the document’s SSL/TLS certificate trust
chain, and to emit the following error:

> javax.net.ssl.SSLHandshakeException:
> sun.security.validator.ValidatorException: PKIX path building failed:
> sun.security.provider.certpath.SunCertPathBuilderException: unable to
> find valid certification path to requested target

Now the behavior for the Web-based checker service is that by default it
no longers examines a document’s SSL/TLS certificate trust chain, which
eliminites the above error.

To instead force the checker service to examine the SSL/TLS certificate
trust chain for all documents it checks that have https URLs, you must
start the service with the `--promiscuous-ssl=off` option specified:

    ./build/build.py --promiscuous-ssl=off run