Add option to check error pages (non-200 response)
This adds a new `checkerrorpages` query parameter to the checker HTTP API, along
with a **check error pages** option in the checker Web UI (under **Options**).
If set, the checker will retrieve a document and check it even if the response
from the server is a 404 or other non-200 status. Otherwise the checker behaves
as it always has: It doesn’t retrieve the document but instead emits a message
indicating the document was not retrieved, along with the HTTP status code.

Documented at https://github.com/validator/validator/wiki/Service:-Common-parameters#checkerrorpages

Fixes #413. Thanks @koalatux