Collapse whitespace before doing langdetect
The raw text content of the relevant element nodes in a document ends up
including a lot of line breaks—mostly just inter-element whitespace.
Feeding that as-is to the language detector seems to significantly
increase the probability of the language detector misidentifying the
language of the content. So this change causes the all the whitespace to
be collapsed in the text content that gets fed to the language detector.