Make our HTML spec parser slightly more robust
The following change to the W3C HTML copy broke the checker build—

https://goo.gl/a6QbHk

That change introduced a (unnecessary and unexpected) `p` element within
a `dt` in the element-summary table for the `footer` element.

So this change updates our HTML spec state-machange parser to just
ignore `p` elements when it is parsing, and just remain in whatever state
it was in before it encountered the `p` element (and consume characters
in that state, if that state is one in which characters get consumed).