Catch+drop ResourceNotRetrievableException in CLI
When the vnu.jar CLI is used for checking an HTTP URL and the Web server
responds with a 404, an uncaught ResourceNotRetrievableException and
stack trace were being emitted (in addition to an actual error message).
This change suppresses the ResourceNotRetrievableException and stack trace.

Fixes https://github.com/validator/validator/issues/460 Thanks @jaimeiniesta