Fix CLI bug misidentifying lang if multiple files
If you give the vnu.jar CLI checker more than 1 documents to check and
the content of any of the 1+N documents is not in the same language as
the 1st, the checker could end up emitting an incorrect warning about
the detected language of the 1+Nth document.

This change fixes that bug. The cause was simply that before the 1+Nth
document check, the code was not clearing a StringBuilder it uses for
storing the document content. So when checking the 2nd document in a
list, that StringBuilder still contained the content of the 1st
document. And when checking the 3rd document in a list, it contained the
content of the 1st document *and* the 2nd document, etc.

Fixes https://github.com/validator/validator/issues/463 Thanks @Fjan