Make “Corrupt GZIP trailer” a non-error
Some Web servers seem to use broken gzip implementations that cause gzipped
content which, despite decompressing successfully, ends up with a gzip trailer
that’s truncated or non-conforming in some way.

Browsers handle such corrupted-trailer gzip content just fine, without even
emitting any errors about it to the console. And the `gzip` tool—although it
will report “unexpected end of file” or something for such gzipped files—will
decompress them successfully anyway.

In fact java.util.zip.GZIPInputStream seems to be about the only processor that
fails to handle such files gracefully—it throws an IOException.

So this change causes the checker code to catch that IOException in the right
place and then just drop it on the floor.