(Re)hide sectioning roots in “Structural outline”
0789dc8 caused headings for sectioning roots to be shown in both the
“Heading-level outline” and “Structural outline” output. It’s correct to
show them in the “Heading-level outline”, but it’s not correct to show
them in the “Structural outline”, which attempts to show an outline view
based on the HTML outline algorithm.

So this change restores the “Structural outline” view to conformance
with the HTML spec, by not including headings for sectioning roots in
the outline, as the spec requires.

However, the “Heading-level outline” (which isn’t intended to conform to
the HTML outline algorithm) continues to show those headings.