Add `filterurl` param to Web checker/network API
To filter out (drop/ignore/suppress) errors and warnings you don’t care
about/don’t want to see/don’t consider to be a problem, use the
`filterurl` param to specify the URL for a filter file. Each line of the
file contains either a regular expression or starts with "#" to indicate
the line is a comment. Any error message or warning message that matches
a regular expression in the file is filtered out (dropped/suppressed).