Skip language detection for figcaption elements
In documents that are largely just comprised of images/figures with
minimal caption text, the language detector may misidentify the language
of the document — especially if the caption text is repetitive.

So skipping the content of figcaption elements when doing language
detection helps to mitigate the misidentification risk.

Fixes https://github.com/validator/validator/issues/530 Thanks @goetzfilm