Suppress warning about CSP child-src directive
In the W3C CSP2 Recommendation the child-src directive is the normative way to
govern the creation of nested browsing contexts, and the frame-src directive is
deprecated, with a note telling readers to use child-src instead.

But in the current CSP spec, the CSP3 Editor’s Draft, the frame-src directive is
“undeprecated” and it’s the child-src directive that is instead deprecated, with
a note telling readers to use frame-src and worker-src instead.

So Salvation appropriately emits a warning saying that the latest CSP spec
deprecates child-src. But it’s not clear what the current state of browser
support is related to this, so to prevent confusion among checker users, it
seems prudent to suppress that warning for now until it’s clear that it’s
recommending what web-developer/authors should actually be using at this point.