Make vnu.jar exit 0 if all errors filtered
This change makes the vnu.jar command-line checker exit 0 if errors were found
but were all filtered out with the `--filterfile` or `--filterpattern` option.

Prior to this change, the checker would exit 1 if any errors were found, even if
the `--filterfile` or `--filterpattern` option were used to filter them all out.

If the `--Werror` option is specified, then for any warnings, this change causes
the same behavior as described above.