Give Web UI/network API ability to check CSS files
This change enables the Web UI and network API to be used for checking CSS
documents: by address/URL, by file upload, and by direct input in the textarea
part of the Web UI — as well as by POSTing CSS content with the network API.

The CSS-document-checking support is implemented on the backend by first taking
the CSS content and packing it into an HTML wrapper, then running the result
through the HTML parser, and then using some special-case handling to deal with
getting correct line-numbering and column-numbering errors, and to unpack the
original CSS source from the HTML wrapper in order to extract portions of it for
highlighting and for linking to from the displayed error message.

The implementation of this CSS-checking support is experimental and may need to
be reverted if the performance impact is too high or if it ends up causing any
unwelcome side effects on the core HTML-checking support.