Enable command-line checker to check CSS files
This change enables the command-line checker to be used to check CSS documents.

The CSS-document-checking support is implemented in the code by first taking
the CSS content and packing it into an HTML wrapper, then running the result
through the HTML parser, and then using some special-case handling to deal with
getting correct line-numbering and column-numbering errors, and to unpack the
original CSS source from the HTML wrapper.

The implementation of this CSS-checking support is experimental and may need to
be reverted if it ends up causing any unwelcome side effects on the core
HTML-checking support.