Disallow nested interactive ARIA roles
For any element with a `role` attribute whose value is defined by the
ARIA in HTML spec as one whose purpose is to mark up interactive
content, this change causes it to be disallowed as a descendant of any
element with `role=button` or `role=link`.

For any elements that the HTML spec defines as interactive content, this
change causes it to be disallowed  as a descendant of any element with
`role=button` or `role=link`.

Fixes https://github.com/validator/validator/issues/422
Thanks @stevefaulkner and @MichielBijl.