Add 'synchronized' to SvgPathData appendToContext
This change is an attempt to fix an NPE that gets thrown occasionally (every few
weeks/months) by the W3C instance of the checker service. If the logs show the
URL for a document that was being checked at the time, I’ve never been able to
reproduce the problem when testing locally with the same URL. Given that, seems
it might be a concurrency/thread-safety issue (the W3C service handles a ton of
concurrent requests...) — so this change attempts to deal with it by adding
`synchronized` to the method the NPE stack trace indicates is the trigger.