Use setOptimizationLevel(-1) when calling Rhino
This change uses `setOptimizationLevel(-1)` for the Context object when
calling the Rhino code to do syntax-checking on JavaScript in `script`
elements. The effect is that it prevents Rhino from compiling the
JavaScript into bytecode and sometimes hitting the 64K limit (as seems
to happen when checking facebook.com, for example).