Allow “async” and “await” in script contents
This change causes instances of the `async` and `await` keywords in
`script` contents to not result in errors — which they otherwise would
due to the fact that the Rhino parser doesn’t recognize them. This
change handles them by doing some preprocessing before passing the
`script` contents to Rhino for parsing.