Allow “const” in scripts wherever “var” is allowed
This change causes `const` instances to be allowed in `script` contents
wherever `var` is allowed, without resulting in errors — which it
otherwise would due to the fact that the Rhino parser/compiler doesn’t
handle `const` in some cases where it’s valid (e.g., for-in/for-of
statements). This change deals with it by doing some preprocessing
before passing the `script` contents to Rhino for parsing+compiling.