Replace Rhino with Acorn for JS syntax checking
This change removes all Rhino dependencies from the code and replaces
them with calls to a new `nu.validator.javascript.JavaScriptParser`
class, which relies on `ScriptEngineManager` to instead load and use the
(written-in-JavaScript) Acorn parser https://github.com/acornjs/acorn.

This effects checks of the following contents:

* `script` elements
* `on*` event-handler attributes (`onclick`, etc.)
* `pattern` attribute
* `javascript:`-protocol URLs (e.g., `<a href>` values)

So the code now performs syntax-checking of those using Acorn to parse
the contents, and to then pass on error messages from Acorn about any
JavaScript syntax (parse) errors it finds in the checked contents.

In contrast to Rhino, which lacks parsing support for many ES6 language
features (see https://mozilla.github.io/rhino/compat/engines.html),
Acorn provides full parsing support for:

* All language features of ES6 (2015), ES7 (2016), and ES8 (2017) too
* A subset of Stage 4 (finished) features for ES9 (2018)
  http://2ality.com/2017/02/ecmascript-2018.html#the-features-of-es2018,
  including the new regular-expression features

However, unlike the Java code of Rhino, which is (apparently)
thread-safe, the replacement Java code relies on Nashorn, which is
(apparently) not thread-safe. So it’s possible that this initial
implementation of the Acorn integration may prove to have some
unexpected buggy behavior in practice once it’s exercised under
real-world-level concurrency/load by the W3C service.

So after this code is deployed for the W3C service, if any concrete
problems/effects related to lack of thread safety do actually become
apparent, it will need to be further refined in some follow-up changes.