End the script-checking experiment
This change retains checking of the `pattern` attribute but drops all
other JavaScript syntax checking — specifically of these contents:

* `script` elements
* `on*` event-handler attributes (`onclick`, etc.)
* `javascript:`-protocol URLs (e.g., `<a href>` values)

The reason for dropping those checks is that examining the checker
behavior in Java profiling tools indicates there’s very high runtime
overhead/cost in using the (written-in-JavaScript) Acorn parser, loaded
through Nashorn and `ScriptEngineManager`, to do the parsing/checking.

Specifically, using that mechanism appears to add another several
hundred milliseconds to the time the checker takes to process any
document that contains `script` elements or event-handler attributes.

That additional cost might be acceptable if there were evidence of some
strong benefit for users. But in practice, it seems there clearly isn’t;
specifically, use-counter statistics collected from the W3C checker
indicate that only about 1% of documents it checks turn out to have
any JavaScript syntax errors.

So it seems very wasteful to significantly slow down checking for the
99% of users with documents that don’t have JavaScript errors just to
potentially help the 1% of users with documents that do have errors.

Also, looking at a sample of the documents for which the checker does
find JavaScript syntax errors indicates that those documents almost
always already have dozens of other (HTML and CSS) problems — so in
practice for the users looking at the checker results for those
documents, the messages about JavaScript syntax errors aren’t likely to
be acted on (those documents are just too big of a mess already).

Still, we can consider once again restoring JavaScript syntax checking
for contents other than just the `pattern` attribute if/when some
performant mechanism for doing it ever becomes available — with the
additional requirement that it also needs to have parsing support for
all features in whatever the current EcmaScript spec happens to be.