Use 'new RegExpFactory.compile()' instead of Acorn
This change replaces the use of Acorn for `pattern`-attribute checking
with (native-Java) Nashorn `new RegExpFactory.compile()`. It turns out
the `pattern` attribute is used widely enough that Java profiling tools
indicate that in practice there’s still a measurable performance penalty
to loading and using Acorn to check it.

So this change also removes the final remaining dependency on Acorn.

Unfortunately, unlike Acorn, the Nashorn `new RegExpFactory.compile()`
replacement lacks support for checking with the `u` flag enabled — but
support for that flag is necessary for full conformance with the
requirements in the HTML spec. So this change also puts our `pattern`
checking out of full conformance with the spec (again). Which means
https://github.com/validator/validator/issues/195 needs to remain open.