Drop gzip filter from war file
The better way to enable compression is to instead use the means that
whatever servlet container already provides for gzipping responses.
For example, in the case of using the war file with Tomcat, edit the
`/usr/local/tomcat/conf/server.xml` file and add a `compression="on"`
attribute to the relevant Connector element.

Closes https://github.com/validator/validator/issues/622