Drop error messages about CSS font-display
This change is a hack to work around the fact the CSS-checking backend
doesn’t yet have support for recognizing the CSS font-display
descriptor. So rather that continuing to emit an error message for it
(which is wrong, because it’s not an error...) this change causes the
checker to filter out (drop/ignore/suppress) error messages about it.

Addresses https://github.com/validator/validator/issues/649