Revert “Increase Jetty idleTimeout to 300000ms”
This change reverts 3b2bde7. Increasing the Jetty idleTimeout interval
didn’t help and anyway it’s become clear that the problem isn’t with
Jetty or the checker code but is instead a problem due to the fact we're
proxying https://checker.html5.org/ and there’s some proxying issue.