Drop CSS row-gap and "align-content: start" errors
This change is a hack to work around the fact the CSS-checking backend
doesn’t yet have support for recognizing the CSS row-gap property
and the "start" value for the align-content property. So rather than
continuing to emit error messages for those (which is wrong, because
they’re not errors...), this change causes the checker to filter out
(drop/ignore/suppress) error messages about those.

Addresses https://github.com/validator/validator/issues/667