Docker: Use binary runtime image instead of jar
This change causes the Dockerfile to use a (Linux) binary runtime image
of the checker (which embeds a subsetted Java JRE), rather than using
the vnu.jar file and needing to include a full JRE in the Docker image.

Using the runtime image of the checker instead of the jar reduces the
Docker image size by almost 200MB (it brings the image size down from
379MB to 185MB). It also prevents any regressions that could be caused
by source/target version conflicts between the built jar and whatever
JRE version would otherwise need to be included in the Docker image.

---

Note: Starting from an Alpine image rather than a Debian image has the
potential to make it (eventually) possible to reduce the Docker image
size even further (by at least 40MB). The current problem with actually
doing it is that Alpine is based on musl but Java needs glibc.

https://github.com/sgerrand/alpine-pkg-glibc provides a general solution
for using glibc-linked binaries with Alpine, but in the specific case of
trying to use it for running the checker from a (jlink-generated) runtime
image, the checker currently fails at runtime with a segfault.