Enable checking standard input as SVG or CSS
This change enables standard input to the command-line checker to be
checked as SVG or CSS — if the --svg or --css options are specified.

Otherwise without this change, even when the --svg or -css options are
specified, any standard input is assumed to be an HTML document.

Fixes https://github.com/validator/validator/issues/808
Thanks @speedyhoon