Build: Run the checker from vnu.jar in classpath
This change causes the "run" target of the build script to use the built
vnu.jar in the classpath to find the needed runtime classes. Without
this change, the build otherwise uses all the jar files in the
dependencies subdirectory to get the needed runtime classes. Since the
built vnu.jar is what we’re distributing, it’s better to actually use
that for the "run" target too. This has the advantage of also exposing
any runtime failure with the built jar file (which would otherwise be
masked if we ran from the jars in the dependencies subdirectory).