Warn when outline can’t be shown due to exception
This change causes the checker to emit a warning when the user has
selected the “Show outline” option but the outline can’t be shown due to
an exception being thrown (and caught internally).

Prior to https://github.com/validator/validator/commit/10a4c52, we had
exposed exceptions to users as “Oops. That was not supposed to happen”
errors — which made it obvious that something went wrong, and so the
outline could not be displayed.

So, without this change to add back an indication to the user that
something went wrong, the user has no idea why the outline did not
display as expected.

Fixes https://github.com/validator/validator/issues/1011