Make System.setProperties ThreadLocal in servlet
This change causes calls to System.setProperty in the servlet code to
be ThreadLocal. This is necessary in the current design because we’re
using System.setProperty in the code which checks the value of the
srcset attribute, and doing that is apparently not thread-safe.

This is a workaround, intended to be temporary until we can figure out
how to change the design to not use System.setProperty for this to begin
with — and instead use some other mechanism that is actually thread-safe.

Related: https://github.com/validator/validator/issues/1099