build: raise CalledProcessError for subprocesses
`build.py` does not always check the exit code of the subprocesses it
calls via `runCmd()`.  This makes it difficult for callers to determine
whether the requested operation succeeded (e.g. `build.py update` or
`build.py docker-build`) and cause operations to silently proceed
despite errors (e.g. `build.py all` runs the validator even if update
failed).

Change `runCmd()` to use `subprocess.check_call()` instead of
`subprocess.call()` so that `CalledProcessError` will be raised if the
subprocess exits with non-zero status.  Remove conditional exit on
`runCmd()` result, since `CalledProcessError` is now raised in this case
and will cause the process to exit with a non-zero code due to an
unhandled exception.

Fixes: #1737
Signed-off-by: Kevin Locke <kevin@kevinlocke.name>