Refine error handling for source-code emitter
When nu.validator.source.SourceCode.emitSource is called, it sometimes
ends up with nu.validator.json.Serializer.charactersImpl throwing a
java.lang.ArrayIndexOutOfBoundsException. It’s not clear why it happens
nor what could be done usefully when it does — so for now, we just catch
the exception and eat it and move on. It seems like the result in this
case is either that the source code doesn’t end up getting emitted at
all, or else it gets truncated. Or something.