(Re)refine error handling for source-code emitter
When nu.validator.source.SourceCode.emitSource is called, it sometimes
ends up with nu.validator.json.Serializer.charactersImpl throwing a
java.lang.ArrayIndexOutOfBoundsException. It’s not clear what causes it
but it’s reproducible with certain URLs — for example, http://lgubiz.net
and http://bombommat.com. So for now, to recover in that case, we just
catch the exception and eat it and move on. The observable result for
the end user is: the emitted source code gets truncated at some point.