Move vnu version number to results pages of checker web form
The W3C https://validator.w3.org/nu/ start page appears to be getting
aggressively cached by some network/intermediate HTTP cache, and a
consequence of that is, the Version indicator in the footer doesn’t
accurately show the version number for the vnu version that’s actually
been deployed there — but instead, the version number of whatever was
deployed the last time the page got cached. (For example, it currently
shows Version: 25.9.25, despite the fact that what’s actually deployed
there currently is v25.10.14.)

So this change moves the version number up closed to the top of the
page, appended to the “Showing results for…” line that shows up in all
results pages after a document is actually checked. Those results pages
never end up getting cached, so they will always accurately show the
correct version number for the vnu version currently deployed.)